show "Next up" in related.
